+++ source_gen/source_gen/source_gen/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 - Add `throwOnUnresolved` configuration to the `GeneratorForAnnotation`
   constructor.
+- Rename `InvalidGenerationSourceError` to `InvalidGenerationSource`. Change
+  from a subtype of `Error` to a subtype of `Exception`. This may be breaking if
+  a builder relies on a `on Exception catch` to ignore this error.
 
 ## 1.4.0